refactor: move loader validation helpers
