+++ src/loaders/baseManifestItemLoader.js
@@ -18,7 +18,7 @@ import { validateDependencies } from '../utils/dependencyUtils.js';
 import { storeItemInRegistry } from './helpers/registryStoreUtils.js';
 import { extractValidFilenames } from '../utils/filenameUtils.js';
 import { summarizeSettledResults } from './helpers/resultsSummary.js';
-import { validateLoadItemsParams } from '../utils/loadersValidationHelpers.js';
+import { validateLoadItemsParams } from './helpers/validationHelpers.js';
 import { processFileWrapper } from './helpers/fileProcessing.js';
 
 // --- Add LoadItemsResult typedef here for clarity ---
